rename: store 도메인 폴더 정리
