+++ src/main/java/yteamserver/domain/bookmark/BookmarkStore.java
@@ -1,7 +1,7 @@
 package yteamserver.domain.bookmark;
 
 import yteamserver.domain.common.BaseEntity;
-import yteamserver.domain.store.repository.Store;
+import yteamserver.domain.store.domain.Store;
 import yteamserver.domain.users.repository.Users;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
