+++ src/main/java/yteamserver/domain/video/application/VideoService.java
@@ -3,8 +3,8 @@ package yteamserver.domain.video.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import yteamserver.domain.store.repository.Store;
-import yteamserver.domain.store.repository.StoreRepository;
+import yteamserver.domain.store.domain.Store;
+import yteamserver.domain.store.domain.repository.StoreRepository;
 import yteamserver.domain.users.repository.UserRepository;
 import yteamserver.domain.users.repository.Users;
 import yteamserver.domain.video.domain.Video;
