+++ src/main/java/yteamserver/domain/video/domain/Video.java
@@ -2,7 +2,7 @@ package yteamserver.domain.video.domain;
 
 import lombok.Builder;
 import yteamserver.domain.common.BaseEntity;
-import yteamserver.domain.store.repository.Store;
+import yteamserver.domain.store.domain.Store;
 import yteamserver.domain.users.repository.Users;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
